refacrtor: AttendanceHistories 리팩토링 진행
함수명, 매개변수 명을 보다 직관적으로 변경하였고, for문을 stream으로 변경하였음.

+++ src/main/java/attendance/domain/CrewStatus.java
@@ -23,7 +23,7 @@ public enum CrewStatus {
         return name;
     }
 
-    public static CrewStatus calculateCrewStatus(Map<AttendanceType, Integer> attendanceResult) {
+    public static CrewStatus calculateCrewStatus(Map<AttendanceType, Long> attendanceResult) {
         int validateValue = 0;
         validateValue += attendanceResult.get(ABSENCE);
         validateValue += attendanceResult.get(LATE) / 3;
